+++ src/components/Switch/Switch.tsx
@@ -3,7 +3,12 @@ import styled from 'styled-components';
 import colors from '../../theme/colors';
 import SwitchWrapper from './SwitchWrapper';
 import Label from '../Label';
-import { AnalyticsProps } from '../../analytics';
+import {
+  AnalyticsComponent,
+  AnalyticsProps,
+  AnalyticsEvents,
+  AnalyticsComponentType,
+} from '../../analytics';
 
 type SwitchProps = {
   checked: boolean;
@@ -68,7 +73,15 @@ const Slider = styled.span`
   }
 `;
 
-const Switch = ({ id, checked, label, disabled = false, onChange, className }: SwitchProps) => {
+const Switch = ({
+  id,
+  checked,
+  label,
+  disabled = false,
+  onChange,
+  className,
+  tag = '',
+}: SwitchProps) => {
   const onToggleChange = (e: ChangeEvent<HTMLInputElement>) => {
     if (e.target) {
       onChange(e.target.checked);
@@ -76,19 +89,37 @@ const Switch = ({ id, checked, label, disabled = false, onChange, className }: S
   };
 
   return (
-    <SwitchWrapper className={className}>
-      <MarginLeftLabel disabled={disabled} htmlFor={id}>
-        <Span>{label}</Span>
-        <Input
-          type="checkbox"
-          id={id}
-          disabled={disabled}
-          checked={checked}
-          onChange={onToggleChange}
-        />
-        <Slider />
-      </MarginLeftLabel>
-    </SwitchWrapper>
+    <AnalyticsComponent>
+      {({ pages, tracker, merchantId, ip }) => (
+        <SwitchWrapper className={className}>
+          <MarginLeftLabel disabled={disabled} htmlFor={id}>
+            <Span>{label}</Span>
+            <Input
+              type="checkbox"
+              id={id}
+              disabled={disabled}
+              checked={checked}
+              onChange={(e) => {
+                if (tag) {
+                  tracker.track({
+                    event: e.target.checked
+                      ? AnalyticsEvents.SWITCH_ON
+                      : AnalyticsEvents.SWITCH_OFF,
+                    type: AnalyticsComponentType.SWITCH,
+                    tag,
+                    pages,
+                    merchantId,
+                    ip,
+                  });
+                }
+                onToggleChange(e);
+              }}
+            />
+            <Slider />
+          </MarginLeftLabel>
+        </SwitchWrapper>
+      )}
+    </AnalyticsComponent>
   );
 };